Improving interrupt handler interface. Removing virtual methods in header files.

+++ Cosa/SPI.hh
@@ -60,7 +60,7 @@ public:
      * Interrupt service on data receive in slave mode.
      * @param[in] data received data.
      */
-    virtual void on_interrupt(uint8_t data);
+    virtual void on_interrupt(uint16_t data);
   };
 
   enum Clock {
